+++ Sources/MapboxCoreNavigation/RouteController.swift
@@ -23,7 +23,7 @@ import os.log
 open class RouteController: NSObject {
     public enum DefaultBehavior {
         public static let shouldRerouteFromLocation: Bool = true
-        public static let reroutingManeuverRadius: ReroutingManeuverOffset = .default
+        public static let reroutingManeuverRadius: ReroutingManeuverBuffer = .default
         public static let shouldDiscardLocation: Bool = false
         public static let didArriveAtWaypoint: Bool = true
         public static let shouldPreventReroutesWhenArrivingAtWaypoint: Bool = true